fix(md-editor): light-theme legibility for Crepe inline code + code blocks (cave-nar)
Verified MdEditor/Crepe on light app themes and fixed two real breaks:

- Inline code rendered as an unreadable dark blob: --crepe-color-inline-area
  was mapped to --code-surface, a FIXED dark surface. Remap inline code to
  the app's theme-aware lavender pill (same recipe as cave-chat.css
  ':not(pre) > code').
- Code blocks shipped Crepe's bundled one-dark CodeMirror colors on a pale
  surface with a stray dark gutter chip. Extract the mood-c CodeMirror
  theme from code-editor.tsx into a shared code-editor-theme.ts and inject
  it into Crepe's code-mirror feature; keep blocks on the always-dark
  --code-surface (the app's read-mode convention) with block chrome
  (language picker/tools) re-scoped to fixed light-on-dark inks.

Verified with Playwright screenshots in coven-mode=light AND dark on the
built app: inline code pill readable, code block matches read-mode blocks,
no dark-mode regression. typecheck + build (bundle in budget) +
test:app 555 green.
